refactor(graphql): Split getFields into countries and nationalities queries
