+++ src/resolvers/types/Entry.ts
@@ -1,0 +1,10 @@
+import { Field, ObjectType, Int } from 'type-graphql';
+
+@ObjectType()
+export class Entry {
+  @Field(() => Int)
+  id: number;
+
+  @Field()
+  value: string;
+}
